server/requirements.txt
aiohttp==3.13.2
certifi==2025.11.12
fastapi==0.127.1
langchain_core==1.2.5
langchain_openai==1.1.6
langgraph==1.0.5
lxml==6.0.2
numpy==2.4.0
python-dotenv==1.2.1
scikit_learn==1.8.0
serpapi==0.1.5
uvicorn==0.40.0
